eden: integrate celery for queued tasks
