+++ eden/hosting.py
@@ -4,13 +4,14 @@ import uvicorn
 import GPUtil
 import threading
 import traceback
-from fastapi import FastAPI, BackgroundTasks
+from fastapi import FastAPI
 
 from .log_utils import Colors
 from uvicorn.config import LOGGING_CONFIG
 from .utils import parse_for_taking_request, write_json, make_filename_and_id, get_filename_from_id, load_json
 from .datatypes import Image
 from .models import Credentials
+from .threaded_server import ThreadedServer
 
 
 '''
@@ -20,7 +21,7 @@ $ celery --app={module}.{method} worker
 from celery import Celery
 
 
-def run_celery_app(app, loglevel = 'INFO', num_workers = 1):
+def run_celery_app(app, loglevel = 'INFO', num_workers = 10):
     """runs a Celery() instance
 
     Args:
@@ -30,8 +31,11 @@ def run_celery_app(app, loglevel = 'INFO', num_workers = 1):
     argv = [
         'worker',
         f'--loglevel={loglevel}',
-        f'--concurrency={num_workers}'
+        f'--concurrency={num_workers}',
+        '-E'
     ]
+
+    print('Running with args: ', argv)
     app.worker_main(argv)
 
 
@@ -42,7 +46,7 @@ def host_block(block,  port = 8080, results_dir = 'results'):
     https://testdriven.io/blog/fastapi-and-celery/
 
     watch this celery worker live with:
-    $ celery --broker="redis://localhost:6379"  flower --port=6060
+    $ celery --broker="redis://localhost:6379" flower --port=6060 
     '''
     celery_app = Celery(__name__)
 
@@ -91,7 +95,7 @@ def host_block(block,  port = 8080, results_dir = 'results'):
             }
 
     @app.post('/run')
-    def start_run(args: block.data_model, background_tasks: BackgroundTasks):
+    def start_run(args: block.data_model):
 
         filename, token = make_filename_and_id(results_dir = results_dir, username = args.username)
 
@@ -114,7 +118,7 @@ def host_block(block,  port = 8080, results_dir = 'results'):
             return status
         else:
             gpu_id = available_gpu_ids[0]
-            background_tasks.add_task(run, args = args, filename =filename, gpu_id = gpu_id)
+            run.delay(args = dict(args), filename = filename, gpu_id = gpu_id)
 
             status = {
                 'status': 'started',
@@ -157,44 +161,15 @@ def host_block(block,  port = 8080, results_dir = 'results'):
     LOGGING_CONFIG["formatters"]["access"]["fmt"] = "[" + Colors.CYAN+ "EDEN" +Colors.END+ "] %(levelprefix)s %(client_addr)s - '%(request_line)s' %(status_code)s"
 
 
-    '''
-    hacks below
-    '''
-
-    import contextlib
-    import time
-    import threading
-    import uvicorn
-    
-    
-    class Server(uvicorn.Server):
-        '''
-        Refer issue: 
-            https://github.com/encode/uvicorn/issues/742#issuecomment-674411676
-        '''
-        def install_signal_handlers(self):
-            pass
-
-        @contextlib.contextmanager
-        def run_in_thread(self):
-            thread = threading.Thread(target=self.run)
-            thread.start()
-            try:
-                while not self.started:
-                    time.sleep(1e-3)
-                yield
-            finally:
-                self.should_exit = True
-                thread.join()
-
     config = uvicorn.config.Config(app = app, port=port)
-    server = Server(config = config)
+    server = ThreadedServer(config = config)
 
+    ## context starts fastAPI stuff
     with server.run_in_thread():
         message = "[" + Colors.CYAN+ "EDEN" +Colors.END+ "]" + " Initializing celery worker on: " + "redis://localhost:6379"
         print(message)
+        ## starts celery app
         run_celery_app(celery_app)
 
     message = "[" + Colors.CYAN+ "EDEN" +Colors.END+ "]" + " Stopped"
     print(message)
-
